pyproject.toml
[project]
name = "character"
version = "0.1.0"
description = "汎用キャラクター生成システム"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"
